added tests for lexer, minor bug fixes

+++ compiler/lexer/lexer.h
@@ -5,16 +5,12 @@
 #include <ctype.h>
 #include <stdlib.h>
 #include <string.h>
-//#define  isbdigit(x) (x == '0' || x == '1')
-//#define  isodigit(x) (isdigit(x) && x != '8' && x != '9')
 #define  ispotentialend(l) (l->type == T_ID || l->type == T_STR || \
             l->type == T_INT64 || l->type == T_FLOAT64 || l->type == T_BREAK || \
             l->type == T_CONT || l->type == T_RPAR || l->type == T_RSQB || \
             l->type == T_RBRC || l->type == T_UNDEF || l->type == T_BOOL)
 
-//inline int isbdigit(char c) {
-//    return c == '0' || c == '1';
-//}
+
 
 /*
 an identifier
@@ -29,7 +25,7 @@ typedef struct {
     char c;
     Token type;
     char *value;
-    int val_len;
+    int64_t val_len;
     int line;
 } Lexer;
